♻️ refactor: SyncData에 TaskHandler 적용(2시간 주기), DIP 리팩토링
