+++ src/main/java/com/project/cheerha/domain/datasync/controller/DataSyncController.java
@@ -1,5 +1,6 @@
-package com.project.cheerha.domain.elasticsearch.data.mysql;
+package com.project.cheerha.domain.datasync.controller;
 
+import com.project.cheerha.domain.datasync.service.DataSyncService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,17 +8,17 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-public class MysqlToElasticsearchSyncController {
+public class DataSyncController {
 
-    private final MysqlToElasticsearchSyncService mysqlToElasticsearchService;
+    private final DataSyncService dataSyncService;
 
     /**
      * 사용자가 원할 때 데이터를 동기화할 수 있는 엔드포인트입니다.
-     * 이 엔드포인트는 POST 요청을 받아 MySQL에서 Elasticsearch로 데이터를 동기화하는 작업을 시작합니다.
+     * 이 엔드포인트는 POST 요청을 받아 데이터를 동기화하는 작업을 시작합니다.
      */
     @PostMapping("/syncData")
     public ResponseEntity<String> startSync() {
-        mysqlToElasticsearchService.syncDataToElasticsearch();
+        dataSyncService.syncData();
         return ResponseEntity.ok("데이터 동기화가 시작되었습니다.");
     }
 }
