+++ src/main/java/com/project/cheerha/domain/datasync/service/DataSync.java
@@ -1,0 +1,6 @@
+package com.project.cheerha.domain.datasync.service;
+
+public interface DataSync {
+
+    void sync();
+}
